Battle Refactor
Pulled battle logic back out of the Unit class as it needed to reference too much information stored in the Battle class.

Also back to using tactics as the complexity of how to use abilities will be too high to take everything into account on the front end, best to just allow the player to have the flexability. Will add more depth anyway.